% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anolis.R
\docType{data}
\name{Anolis}
\alias{Anolis}
\title{Datos de Lagartos Anolis de Torres de Transectos}
\format{
Un data frame con 503 filas y 15 columnas:
\describe{
  \item{STUDY}{Nombre del estudio (marca y recaptura). Carácter.}
  \item{Survey_Site}{Sitio de muestreo. Carácter.}
  \item{LOCATION}{Área de muestreo. Carácter.}
  \item{TIME}{Momento de la recolección de datos (mañana, mediodía, tarde). Carácter.}
  \item{DATE}{Fecha del muestreo. Tipo fecha o carácter.}
  \item{SEASON}{Época del año. Carácter.}
  \item{SPECIES}{Especies de lagarto Anolis observadas. Carácter.}
  \item{SEX_AGE}{Sexo o edad del lagarto (Female, Juvenil, Male). Carácter. En juveniles no se puede determinar el sexo.}
  \item{HEIGHT}{Altura sobre el nivel del suelo donde fue observado el lagarto (en metros). Numérico (doble).}
  \item{DISTANCE_FROM_CENTERLINE}{Distancia horizontal desde la línea central (en metros). Numérico (doble).}
  \item{PERCH_SUBSTRATE}{Tipo de sustrato o percha donde se encuentra el lagarto. Carácter.}
  \item{PERCH_DIAMETER}{Diámetro del sustrato o percha (en centímetros). Numérico (doble).}
  \item{WEIGHT}{Peso del lagarto (en gramos). Numérico (doble).}
  \item{SVL}{Longitud Hocico-Cloaca (Snout-Vent Length) del lagarto en milímetros. Numérico (doble).}
  \item{TAIL}{Largo de la cola del lagarto en milímetros. Numérico (doble).}
}
}
\source{
Datos del programa Big Grid, Estacion Experimental de El Verde,
Bosque Nacional El Yunque, Puerto Rico.
}
\usage{
data(Anolis)
}
\description{
Datos de distribución vertical y abundancia relativa de especies de lagartos
Anolis recolectados mediante transectos en torres entre 1989 y 1992.
}
\details{
Los transectos fueron conducidos caminando lentamente hacia arriba en cada
torre y registrando información de cada lagarto observado: especie, sexo/edad,
altura de la percha, sustrato de percha, diámetro de percha y distancia desde
la línea central. Se realizaron tres transectos replicados durante cada estación,
con un muestreo en cada uno de tres momentos del día (mañana, mediodía y
tarde).

Los datos permiten calcular la distancia promedio de avistamiento para cada
especie, que se utiliza en el cálculo final de abundancia.
}
\examples{
data(Anolis)
head(Anolis)
library(ggplot2)
ggplot(Anolis, aes(x = HEIGHT, y = SVL, color = SPECIES)) +
  geom_point() +
  facet_wrap(~ SEASON) +
  theme_minimal()

}
\keyword{datasets}
