% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caladeniavalida.R
\docType{data}
\name{caladeniavalida}
\alias{caladeniavalida}
\title{Orquídea Endangered Caladenia valida}
\format{
Un data frame con 164 filas y 15 columnas:
\describe{
  \item{Population}{numeric: número de población para identificación en campo}
  \item{Year}{numeric: año de observación}
  \item{Plant_num}{numeric: número único de identificación de cada planta}
  \item{OH}{numeric: altura total de la planta desde el suelo (mm)}
  \item{OD}{numeric: profundidad total de la planta (mm)}
  \item{OW}{numeric: ancho total de la planta (mm)}
  \item{DSL}{numeric: largo del sépalo dorsal (mm)}
  \item{DCL}{numeric: largo de la protuberancia dorsal (mm)}
  \item{LSL}{numeric: largo del sépalo lateral (mm)}
  \item{LCL}{numeric: largo de la protuberancia lateral (mm)}
  \item{LSW}{numeric: ancho del sépalo lateral (mm)}
  \item{PL}{numeric: largo del pétalo (mm)}
  \item{LL}{numeric: largo del labelo (mm)}
  \item{LW}{numeric: ancho del labelo (mm)}
  \item{Fruit_not}{numeric: presencia de fruto/cápsula (1=presente, 0=ausente, NA=desconocido)}
}
}
\source{
Datos recolectados por Andrew Bould, Isla Raymond,
Victoria, Australia. Mas de 200 especimenes monitoreados
por 15 anos consecutivos.
}
\usage{
data(caladeniavalida)
}
\description{
Datos de la orquídea terrestre en peligro de extinción Caladenia valida de la Isla Raymond,
Victoria, Australia. Los datos fueron recolectados durante 15 años por Andrew Bould.
Incluyen medidas morfológicas de flores, hojas y presencia de frutos.
}
\examples{
head(caladeniavalida)

}
\references{
Dixon K & Tremblay RL. 2009.
  Biology and natural history of Caladenia.
  Australian Journal of Botany 57: 247-258.

  Tremblay RL, Pérez ME, Larcombe M, Brown A, Quarmby J, Bickerton D,
  French G & Bould A. 2009. Population dynamic of Caladenia:
  Bayesian estimates of transition and extinction probabilities.
  Australian Journal of Botany 57: 351-360.

  Tremblay RL, Pérez ME, Larcombe M, Brown A, Quarmby J, Bickerton D,
  French G & Bould A. 2009. Dormancy in Caladenia:
  A Bayesian approach to evaluating latency.
  Australian Journal of Botany 57: 340-350.
}
\keyword{datasets}
