% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crecimiento_domestico_bruto.R
\docType{data}
\name{Crecimiento_domestico_bruto}
\alias{Crecimiento_domestico_bruto}
\title{Crecimiento_domestico_bruto}
\format{El archivo incluye 13 columnas con 28 filas
\describe{
 \item{Pais}{Identificación del pais}
 \item{X1990, X2000, X2006 al X2015}{El crecimiento domestico bruto anual en los paises mencionados en estos años }
}}
\usage{
data(Crecimiento_domestico_bruto)
}
\description{
Los datos son del siguiente sitio, y la metadata que explica los datos
}
\details{
Code	Indicator
NY.GDP.PCAP.KD.ZG

GDP per capita growth (annual %), los datos son del
World Bank national accounts data, and OECD National Accounts data files.
}
\examples{
head(Crecimiento_domestico_bruto)
}
\keyword{datasets}
