% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internet2.R
\docType{data}
\name{Internet2}
\alias{Internet2}
\title{Internet2}
\format{El archivo incluye 4 columnas con 134 filas
\describe{
 \item{Pais}{Paises de latino america}
 \item{SubscripcionesMobiles}{La cantidad de subscripciones de teléfono mobile}
 \item{UsariosInternet}{La cantidad de usuarios de Internet}
 \item{PIBperCapita}{El Producto Interior Bruto per capita}
 \item{AlfabetismoPor}{El porcentaje de alfabetismo}
 \item{UrbanismoPor}{El porcentaje de urbanismo}
 }
 Data del "World Health Organization", supplementado por información de los paises.}
\usage{
data(Internet2)
}
\description{
Internet2
}
\examples{
head(Internet2)
}
\keyword{dataset}
