% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lelto.R
\docType{data}
\name{Lelto}
\alias{Lelto}
\title{Lelto}
\format{An object of class data frame}
\usage{
data(Lelto)
}
\description{
Los datos son de la especies de orquídea epífitas Lepanthes eltoroensis
Los datos fueron recolectado en el area de la vereda de los vientos alizios
en el bosque nacional del Yunque, Puerto Rico
por Dr. Raymond L. Tremblay y sus estudiantes.
}
\details{
Tree: El número del arbol

Trail. La nombre de la vereda

Alt_m: Altitude en metro sobre el nivel del mar

Angulo: El angulo donde se encuentra la mayoria de los individuos sobre el tronco del arbol, va de 0 a 360 grados, donde 0 y 360° es el norte y 180 es el sur.

Rango_Rd: los valores de los ángulos en radianes

Localización: localización de las poblaciones del comienzo de la primer población en la vereda

Inclinación: la inclinación del tereno en ángulo

Plantulas: El cantidad de plantulas en el árbol

Juneniles: El cantidad de juveniles en el árbol

Adulto: El cantidad de adultos en el árbol

T_Num: El cantidad total de orquídeas en el árbol

Lado_vereda: El lado en que se encuentra el árbol en la vereda, derecha (sur) o izquierda (norte)

long: Longitud

lat: Latitude
}
\examples{
head(Lelto)
}
\keyword{datasets}
