% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIB_vs_Alfabetismo.R
\docType{data}
\name{PIB_vs_Alfabetismo}
\alias{PIB_vs_Alfabetismo}
\title{Producto Interno Bruto versus Alfabetismo en América Latina}
\format{
Un data frame con 20 filas y 3 columnas:
\describe{
  \item{Pais}{character: nombre del país de América Latina}
  \item{Alfabetismo}{numeric: porcentaje de alfabetismo (\%)}
  \item{Porciento_PIB}{numeric: porcentaje del PIB dedicado a educación (\%)}
}
}
\source{
Banco Mundial. Indicadores de Desarrollo Mundial.
PIB per capita y tasa de alfabetismo (SE.ADT.LITR.ZS).
\url{https://data.worldbank.org/}
}
\usage{
data(PIB_vs_Alfabetismo)
}
\description{
Datos que relacionan el producto interno bruto (PIB) dedicado a educación
con los niveles de alfabetismo en 20 países de América Latina.
}
\examples{
head(PIB_vs_Alfabetismo)

}
\keyword{datasets}
