% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIBvsSalud.R
\docType{data}
\name{PIB_vs_Salud}
\alias{PIB_vs_Salud}
\title{Producto Interno Bruto en Salud versus Esperanza de Vida}
\format{
Un data frame con 20 filas y 3 columnas:
\describe{
  \item{Pais}{character: nombre del país}
  \item{Porcentaje_PIB_Salud}{numeric: porcentaje del gasto público en salud respecto al PIB (\%)}
  \item{Esperanza_vida}{numeric: esperanza de vida al nacer (años)}
}
}
\source{
globalEDGE, Michigan State University.
Indicadores de gasto en salud y esperanza de vida.
\url{https://globaledge.msu.edu/}
}
\usage{
data(PIB_vs_Salud)
}
\description{
Datos que relacionan el gasto en salud como porcentaje del producto interno bruto (PIB)
con la esperanza de vida al nacer en 20 países.
}
\examples{
head(PIB_vs_Salud)

}
\keyword{datasets}
