% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_PR.R
\docType{data}
\name{Pop_PR}
\alias{Pop_PR}
\title{Población de Puerto Rico por Edad y Sexo}
\format{
Un data frame con 36 filas y 3 columnas:
\describe{
  \item{Sexo}{character: sexo de la población (Hombre/Mujer)}
  \item{Valor}{numeric: número de personas en el cohorte de edad}
  \item{Edad}{character: cohorte de edad de la población}
}
}
\source{
Oficina del Censo de los Estados Unidos, estimaciones de
poblacion de Puerto Rico, 2015.
\url{https://www.census.gov/quickfacts/PR}
}
\usage{
data(Pop_PR)
}
\description{
Datos de población de Puerto Rico por cohorte de edad y sexo del Censo de EE.UU. 2015.
}
\examples{
head(Pop_PR)

}
\keyword{datasets}
