% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Razon_mortandad.R
\docType{data}
\name{Razon_mortandad}
\alias{Razon_mortandad}
\title{Razón de Mortalidad de Menores de Cinco Años}
\format{
Un data frame con 194 filas y 11 columnas:
\describe{
  \item{Pais}{character: nombre del país}
  \item{X2006}{numeric: razón de mortalidad 2006 (por 1,000 nacimientos)}
  \item{X2007}{numeric: razón de mortalidad 2007 (por 1,000 nacimientos)}
  \item{X2008}{numeric: razón de mortalidad 2008 (por 1,000 nacimientos)}
  \item{X2009}{numeric: razón de mortalidad 2009 (por 1,000 nacimientos)}
  \item{X2010}{numeric: razón de mortalidad 2010 (por 1,000 nacimientos)}
  \item{X2011}{numeric: razón de mortalidad 2011 (por 1,000 nacimientos)}
  \item{X2012}{numeric: razón de mortalidad 2012 (por 1,000 nacimientos)}
  \item{X2013}{numeric: razón de mortalidad 2013 (por 1,000 nacimientos)}
  \item{X2014}{numeric: razón de mortalidad 2014 (por 1,000 nacimientos)}
  \item{X2015}{numeric: razón de mortalidad 2015 (por 1,000 nacimientos)}
}
}
\source{
Banco Mundial. Indicador: SH.DYN.MORT (Tasa de mortalidad
de menores de 5 anos por cada 1,000 nacidos vivos).
\url{https://data.worldbank.org/indicator/SH.DYN.MORT}
}
\usage{
data(Razon_mortandad)
}
\description{
Datos de la razón de mortalidad de niños menores de cinco años por cada 1,000 nacimientos
para 194 países entre 2006 y 2015. Los datos provienen del Banco Mundial.
}
\examples{
head(Razon_mortandad)

}
\keyword{datasets}
