% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiroide.R
\docType{data}
\name{Tiroide}
\alias{Tiroide}
\title{Tiroide}
\format{An object of class data frame}
\usage{
data(Tiroide)
}
\description{
Se utilizarán datos sobre la cantidad de
nuevos casos de personas con problema de
la tiroide en EE UU y la mortandad por
cada 100,000 habitantes.
Los datos provienen del National Cancer
Institute; Surveillance, Epidemiology and
End Results Program,
}
\details{
Los datos de las tres columnas son
Año del muestreo,
Los casos nuevos o posteriomente registrado y
la razón de tiroide por 100,000 habitantes.
}
\examples{
head(Tiroide)
}
\references{
seer.cancer.gov
}
\keyword{datasets}
