% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegSampleV1.R
\docType{data}
\name{VegSamplesV1}
\alias{VegSamplesV1}
\title{Datos de Muestreo de Vegetación}
\format{
Un data frame con 60 filas y 19 columnas:
\describe{
  \item{Year}{numeric: año de muestreo}
  \item{Site}{character: identificación del sitio}
  \item{UniversalPlotName}{character: nombre universal del parcela}
  \item{Banded}{numeric: indicador de banding}
  \item{PtCountsum}{numeric: suma de conteo de puntos}
  \item{Avgmaxht}{numeric: altura máxima promedio (m)}
  \item{Avgdens}{numeric: densidad promedio}
  \item{ht.thatch}{numeric: altura de paja}
  \item{S.patens}{numeric: cobertura de Spartina patens (\%)}
  \item{Distichlis}{numeric: cobertura de Distichlis (\%)}
  \item{S.alternifloraShort}{numeric: cobertura de Spartina alterniflora corta (\%)}
  \item{S.alternifloraTall}{numeric: cobertura de Spartina alterniflora alta (\%)}
  \item{Juncus}{numeric: cobertura de Juncus (\%)}
  \item{Bare}{numeric: cobertura desnuda (\%)}
  \item{Other}{numeric: cobertura de otras especies (\%)}
  \item{Phragmites}{numeric: cobertura de Phragmites (\%)}
  \item{Shrub}{numeric: cobertura de arbustos (\%)}
  \item{Tallsedge}{numeric: cobertura de juncos altos (\%)}
  \item{Water}{numeric: cobertura de agua (\%)}
}
}
\source{
Datos usados con permiso de Alain F. Zuur. Disponibles en
\url{https://www.highstat.com/}
}
\usage{
data(VegSamplesV1)
}
\description{
Datos de muestreo de vegetación de 60 sitios. Incluye información sobre
cobertura de especies, altura y densidad de vegetación. Datos utilizados
con permiso de Zuur, Ieno y Elphick (2010). Fuente: highstat.com
}
\examples{
head(VegSamplesV1)

}
\references{
Zuur AF, Ieno EN & Elphick CS. 2010.
  A protocol for data exploration to avoid common statistical problems.
  Methods in Ecology and Evolution 1: 3-14.
}
\keyword{datasets}
