% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASBESTOS_QUEBEC.R
\docType{data}
\name{ASBESTOS_QUEBEC}
\alias{ASBESTOS_QUEBEC}
\title{Asbestos_Quebec}
\format{El archivo incluye 5 columnas con 466 filas
\describe{
 \item{Fecha}{La fecha en año y mes}
 \item{Year}{El año}
 \item{Mes}{El mes del año}
 \item{Temp_Prom_Max}{El promedio de la temperatura maxima en el mes}
 \item{Temp_Prom_Min}{El promedio de la temperatura minima en el mes}
 }}
\usage{
data(ASBESTOS_QUEBEC)
}
\description{
Los datos son temperatura minima y maxima del
pueblo de Asbestos, Quebec, Canada del 1948 al 1987.
}
\examples{
head(ASBESTOS_QUEBEC)
}
\keyword{datasets}
