% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Camas_Hospital.R
\docType{data}
\name{Camas_Hospital}
\alias{Camas_Hospital}
\title{Camas_Hospital}
\format{El archivo incluye 4 columnas con 134 filas
\describe{
 \item{Pais}{Paises del mundo}
 \item{Year}{El año}
 \item{Poblacion}{El tamaño poblacional}
 \item{Camas}{El número de camas por 1,000 personas}
 }
 Data del "World Health Organization", supplementado por información de los paises.}
\usage{
data(Camas_Hospital)
}
\description{
Camas_Hospital
}
\examples{
head(Camas_Hospital)
}
\keyword{datasets}
