% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CypripediumA.R
\docType{data}
\name{CypripediumA}
\alias{CypripediumA}
\title{CypripediumA}
\format{An object of class data frame}
\usage{
data(CypripediumA)
}
\description{
Los datos son de la especies de orquidea terrestre Cypripedium acaule.
Los datos fueron recolectado en el area de North Bay, Ontario, Canada
por Dr. Peter Norsko y sus estudiantes.
}
\details{
Orchid_no: El numero de la planta

Forest_Wetland: Si la planta esta localizada en el bosque o en un humedal

Part_of_clump: Si la planta es parte de grupo de individuos o es solitaria

Total_height_cm: La altura de la planta de la base hasta la flor

Leaf_1_length_mm: El largo de la primera hoja en millimetros.
La planta tiene al maximo dos hojas

Leaf_2_length_mm: El largo de la segunda hoja en millimetros,
 la hoja más cercana al suelo

Dorsal_petal_length_mm: El largo del petalo dorsal en millimetros

Lip_length_mm: El largo del labello en millimetros

Lip_width_mm: El ancho del labello en millimetros
}
\examples{
head(CypripediumA)
}
\keyword{datasets}
