% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edu_Salud_Gastos_GDP.R
\docType{data}
\name{Edu_Salud_Gastos_GDP}
\alias{Edu_Salud_Gastos_GDP}
\title{Edu_Salud_Gastos_GDP}
\format{El archivo incluye 4 columnas con 156 filas
\describe{
 \item{Pais}{El pais}
 \item{Gasto_Salud_percapita}{El gasto aportado a la salud per capita}
 \item{Porc_Ninas_no_escuela}{El porcentaje de niñas que no estan en la escuela}
 \item{Porc_Gastos_Educ}{El porcentaje de gastos del PIB dedicado a la educación}
 }}
\usage{
data(Edu_Salud_Gastos_GDP)
}
\description{
Edu_Salud_Gastos_GDP
}
\examples{
head(Edu_Salud_Gastos_GDP)
}
\keyword{datasets}
