% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Educacion_Ninas.R
\docType{data}
\name{Educacion_Ninas}
\alias{Educacion_Ninas}
\title{Educacion_Ninas}
\format{El archivo incluye 13 columnas con 25 filas
\describe{
 \item{Serie}{Porcentaje de Niña matriculada en pre-primeria, NMPP}
 \item{2010NMPP}{El año 2010,  NMPP}
 \item{2011NMPP}{El año 2011,  NMPP}
 \item{2012NMPP}{El año 2012,  NMPP}
 \item{2013NMPP}{El año 2013,  NMPP}
 \item{2014NMPP}{El año 2014,  NMPP}
 \item{Pais}{Los paises}
 \item{Serie2}{Porcentaje de mujer matriculada en Escuela primaria, PMEP}
 \item{2010PMEP}{El año 2010,  PMEP}
 \item{2011PMEP}{El año 2011,  PMEP}
 \item{2012PMEP}{El año 2012,  PMEP}
 \item{2013PMEP}{El año 2013,  PMEP}
 \item{2014PMEP}{El año 2014,  PMEP}
 }}
\usage{
data(Educacion_Ninas)
}
\description{
Educacion_Ninas
}
\examples{
head(Educacion_Ninas)
}
\keyword{datasets}
