% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKERT_DATA.R
\docType{data}
\name{LIKERT_DATA}
\alias{LIKERT_DATA}
\title{LIKERT_DATA}
\format{El archivo incluye 3 columnas con 30 filas. Representando respuestas de 30 estudiantes a tres preguntas

\describe{
 \item{P1}{Tenemos equipo de simulación adecuado para practicar?}
 \item{P2}{Disponemos de modulos computarizados adecuados para el aprendizaje y la enseñanza?}
 \item{P3}{Están los cursos de laboratorio al día?}
\item{Alternativas de repuestas}{Las respuestas eran alternativas numeradas de 1, completamente en desacuerdo hasta el 5, completamente de acuedo}
 }}
\usage{
data(LIKERT_DATA)
}
\description{
LIKERT_DATA
}
\examples{
head(LIKERT_DATA)
}
\keyword{datasets}
