% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Morelia_Mich_Tmin.R
\docType{data}
\name{MORELIA.MICH.Tmin}
\alias{MORELIA.MICH.Tmin}
\title{MORELIA.MICH.Tmin}
\format{An object of class data frame}
\usage{
data(MORELIA.MICH.Tmin)
}
\description{
Los datos provienen de clicom-mex.cicese.mx.
La temperatura minima de cada día desde Noviembre 1947 a marzo 2012.
}
\details{
Dia: El dia del mes de muestreo

Mes: El mes del muestreo

Year: El año de muestreo

Tmin: La temperatura mínima del día
}
\examples{
head(MORELIA.MICH.Tmin)
}
\keyword{datasets}
