% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartosInfantes.R
\docType{data}
\name{PartosInfantes}
\alias{PartosInfantes}
\title{PartosInfantes}
\format{El archivo incluye 5 columnas con 156 filas, los datos provienen del World Development Indicators, 07_22_2016.
\describe{
 \item{NMI}{Número de Muertes de Infantes, Niños que mueren en el parto o natimuerto}
 \item{NMP}{Número de Muertes por Parto, las madres que mueren al parir}
 \item{GSPC}{El en Salud per Cápita, gasto aportado a la salud por persona}
 \item{Grupo}{Región del mundo}
 \item{Pais}{El nombre del pais}
 }}
\usage{
data(PartosInfantes)
}
\description{
PartosInfantes
}
\examples{
head(PartosInfantes)
}
\keyword{datasets}
