% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBI.R
\docType{data}
\name{PBI}
\alias{PBI}
\title{Producto Bruto Interno}
\format{An object of class data frame}
\usage{
data(PBI)
}
\description{
Los datos son del siguiente sitio,Code	Indicator: NY.GDP.PCAP.KD.ZG
Cada columna representa le crecimiento de un año al otro.
GDP per capita growth (annual %)
}
\details{
Annual percentage growth rate of GDP per capita based on constant local currency.

Aggregates are based on constant 2010 U.S. dollars.

GDP per capita is gross domestic product divided by midyear population.

GDP at purchaser's prices is the sum of gross value added by all resident producers in the economy plus any product taxes and minus any subsidies not included in the value of the products.
It is calculated without making deductions for depreciation of fabricated assets or for depletion and degradation of natural resources.
World Bank national accounts data, and OECD National Accounts data files.
}
\examples{
head(PBI)
}
\keyword{datasets}
