% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIB_vs_Alfabetismo.R
\docType{data}
\name{PIB_vs_Alfabetismo}
\alias{PIB_vs_Alfabetismo}
\title{PIB_vs_Alfabetismo}
\format{El archivo incluye 3 columnas con 20 filas
\describe{
 \item{Pais}{Pais de Latino America}
 \item{Alfabetismo}{El porciento de alfabetismo}
 \item{Porciento_PIB}{El porciento del producto interior dedicado a la educación}
 }}
\usage{
data(PIB_vs_Alfabetismo)
}
\description{
PIB_vs_Alfabetismo
}
\examples{
head(PIB_vs_Alfabetismo)
}
\keyword{datasets}
