% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIBvsSalud.R
\docType{data}
\name{PIB_vs_Salud}
\alias{PIB_vs_Salud}
\title{PIB_vs_Salud}
\format{An object of class data frame}
\usage{
data(PIB_vs_Salud)
}
\description{
Los datos provienen de http://globaledge.msu.edu.
}
\details{
Los datos representan el porcentaje de producto bruto dedicado a la salud por diferentes paises
y un indice de salus que es la esperanza de vida de un niño al nacer.

Pais: Los paises en el archivo de datos

Porcentaje_PIB_Salud: El porcentaje del gasto publico del Producto Interior Bruto dedicado a la salud por el pais

Esperanza_vida: La esperanza de vida de un niño o niña que nace en este pais
}
\examples{
head(PIB_vs_Salud)
}
\keyword{datasets}
