% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_PR.R
\docType{data}
\name{Pop_PR}
\alias{Pop_PR}
\title{Pop_PR}
\format{El archivo incluye 3 columnas con 37 filas
\describe{
 \item{Edad}{El cohorte de edad de la poblacion}
 \item{Sexo}{Variable discreta para Mujer y Hombre}
 \item{Valor}{El tamaño poblacional del cohorte}
 }
Datos de Censo de EEUU de 2015.}
\usage{
data(Pop_PR)
}
\description{
El archivo contiene el tamaño poblacional de Puerto Rico en 2015
}
\examples{
head(Pop_PR)
}
\keyword{datasets}
