% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Razon_mortandad.R
\docType{data}
\name{Razon_mortandad}
\alias{Razon_mortandad}
\title{Razon_mortandad}
\format{An object of class data frame}
\usage{
data(Razon_mortandad)
}
\description{
Los datos por pais entre 2006 y 2015 y representan
sobre la razón de mortandad de niños de menor de 5 años por cada 1,000 nacimiento
}
\details{
Los Datos provienen del "Banco Mundial", "World Bank" en Ingles en la siguiente dirección

data.worldbank.org/indicator/SH.DYN.MORT

Los datos incluyen la razón por de mortandad de niños menor de 5 años cada pais entre los años indicado ariba
}
\examples{
head(Razon_mortandad)
}
\keyword{datasets}
