% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparrowsElphick.R
\docType{data}
\name{SparrowsElphick}
\alias{SparrowsElphick}
\title{SparrowsElphick}
\format{An object of class data frame}
\usage{
data(SparrowsElphick)
}
\description{
SparrowsElphick
}
\examples{
head(SparrowsElphick)
}
\keyword{datasets}
