% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegSampleV1.R
\docType{data}
\name{VegSamplesV1}
\alias{VegSamplesV1}
\title{VegSamplesV1}
\format{An object of class data frame}
\usage{
data(VegSamplesV1)
}
\description{
Con el permiso de Alain F Zuur usamos los datos
}
\details{
Toda la información es exactamente como aparece en el archivo original de Zuur, Ieno y Elphick 2010.
"A protocol for data exploration to avoid common statistical problems" Methods in Ecology and Evolution 1: 3-14.

Estaré solamente explicando el nombre de las columnas que estan relevante a los gráficas producidas.

Sugiero que tambien evaluan el siguiente libro para más ejemplos como tomar en cuenta los supuestos de las pruebas.
"A Beginner's Guide to Data Exploration and Visualisation with R".
por Elena N Ieno y Alain F Zuur. 2015. Publicado por Highland Statistics Ltd.

Los datos fueron bajado del siguiente sitio de Web en al "Appendix S1 (Supporting Information),
www.highstat.com

En el archivo hay 19 columnas y 60 filas
Los datos usados en los ejercicios del libro son limitado a las siguientes columnas
}
\examples{
head(VegSamplesV1)
}
\keyword{datasets}
